Centralize Types and Dependencies
