--- gmail-senders/deps.ts
@@ -1,2 +1,0 @@
-export { join } from "https://deno.land/std@0.168.0/path/mod.ts";
-export { open } from "https://deno.land/x/open@v0.0.5/index.ts";
